pyproject.toml
[project]
name = "vish-backend"
version = "0.0.1"
description = "Backend for vish"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "environs[django]>=14.1.1",
    "markdown>=3.7",
    "mysqlclient>=2.2.7",
]
